Update main.py
Fixed a few things. You should now be able to automate this being run by using cron. The issue was coming from the config technically not being accessible. It works for me on Ubuntu Server. 

Another issue was also with the fact that shifts that end at midnight don't make sense to the GCal API. Shifts that end at midnight will now automatically change to 11:59. 

Also it appears Target changes the XPath for the button that goes to the next week. Weird. Regardless, Fixed that too.